requirements.txt
langchain
langchain-experimental
langchain-anthropic
python-dotenv
opencv-python
scikit-image
scipy
numpy
matplotlib
